pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "FridAI task management frontend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "python-fasthtml>=0.4.5",
    "pyyaml>=6.0",
    "uvicorn==0.30.6",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "httpx>=0.27.2",
    "respx>=0.21.1",
]
